Refactoring components
